+++ actix-http-test/CHANGES.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+- Add `TestServer::{query, squery}()` for issuing HTTP `QUERY` requests (RFC 10008). [#4140]
 - Minimum supported Rust version (MSRV) is now 1.88.
+
+[#4140]: https://github.com/actix/actix-web/pull/4140
 
 ## 3.2.0